pyproject.toml
[project]
name = "paperbrace"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pymupdf>=1.24",
  "sqlite-utils>=3.36",
]

[project.optional-dependencies]
chroma = [
  "chromadb",
  "sentence-transformers",
]
sqlitevec = [
  "sqlite-vec",
  "sentence-transformers",
  "pysqlite3-binary",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
paperbrace = "paperbrace.cli:app"
